Test that all edges between nodes in sampled subgraph are included in the output.
This is sometimes known as node-based sampling, where the entire adjacency list of every node in the subgraph is filtered by the unique set of node IDs in the subgraph.

An alternative technique (not yet currently supported) is edge-based sampling, a configuration where only edges traversed during sampling are retained in the output subgraph.

PiperOrigin-RevId: 454610152