When computing the per-head `values`, project the node and edge inputs separately, similarly to as is done for the `keys`.
This allows us to project the `sender_node_input` values before they are broadcast to the edges, which means that we project each value once per node and head instead of once per edge and head.

Since we usually have more edges than nodes, this change provides a non-negligible performance gain.

Due to changes in some of the attribute names, this change breaks checkpoint compatibility between models before and after this change using the `MultiHeadAttentionConv` layer.

PiperOrigin-RevId: 477559910