Apply the per-head transformation of the `values` _after_ they have been interpolated with the attention weights, as opposed to transforming them before interpolating (the previous behaviour).
This change provides a significant speedup for models with edge inputs, since
before interpolating the transformation is applied to a Tensor with leading
dimension equal to the number of edges, and after interpolating it is only applied
to a Tensor of leading dimension equal to the number of receiver nodes.

Due to changes in some of the attribute names, this change breaks checkpoint compatibility between models before and after this change using the MultiHeadAttentionConv layer.

This change is based on work and code originally written by https://github.com/LiangzeJiang.

PiperOrigin-RevId: 479584207