Modify TF-GNN modeling to cope with auxiliary node sets and edge sets.
`tfgnn.check_homogeneous_graph_tensor()` now ignores aux graph pieces.
The new function `tfgnn.get_homogeneous_node_and_edge_set_name()` does
the same uniqueness check and then and returns the tuple of the unique
non-aux `(node_set_name, edge_set_name)` it has found.

`tfgnn.keras.ConvGNNBuilder` now ignores aux graph pieces and raises
a ValueError if user input tries to force it anyways.

The `*HomGraphUpdate` and heterogeneous `*GraphUpdate` classes of the
model collection are updated accordingly, **except** GraphSAGE, which
is not based on `ConvGNNBuilder` and gets a TODO about that.

By default, a graph piece counts as auxiliary if it matches the
`tfgnn.AUX_GRAPH_PIECE_PATTERN` (with `re.fullmatch()`). The core library
functions get a kwarg to override this. Models can expose that argument
if a need for it is discovered.

PiperOrigin-RevId: 516844475