Refactor `GraphSAGEGraphUpdate` based on `ConvGNNBuilder`, to let it handle auxiliary node sets and edge sets properly, by picking up the changes from commit https://github.com/tensorflow/gnn/commit/cde55bcea3f5b4c4c4bdd9c3fe81585d068797f5
This refactoring adds an error message for a corner case:
If `GraphSAGEGraphUpdate(node_set_names=[...])` is called with
an explicitly specified node set that does not receive convolution
results (i.e., aggregated messages) from any edge set, a ValueError
is raised. Previously, such a node set was silently ignored, against
the literal appearance of the calling code. To fix, simply omit the
offending node set from the list; the resulting model is unchanged.

PiperOrigin-RevId: 529028556