In `tfgnn.add_self_loops(...)`, drop `edge_feature_initializer=`, because the current interface is unsuitable except for features that are not ragged and have dtype `tf.float32`.
`tfgnn.add_self_loops(...)` now works for features of arbitrary dtypes.

We suspect there is no actual usage of the removed callback. If you need to
control the feature values of added edges, please file a feature request.
(We now think a dict of values is easier, if needed at all.)

As an immediate workaround, consider adding an all-true boolean feature on
the edge set. New edges will have it set to false, enabling `tf.where()` for
further adjustments.

PiperOrigin-RevId: 580543607